+++ src/apps/ScopeChangeRequest/Components/AdvancedDocumentSearch/StidDocument.tsx
@@ -4,7 +4,7 @@ import { Icon } from '@equinor/eds-core-react';
 import styled from 'styled-components';
 import { getDocumentById } from '../../api/STID/getDocumentById';
 import { transformIsoDate } from '../Workflow/Utils/dateFormatting';
-import { useInfiniteCachedQuery } from '../../hooks/React-Query/useInfiniteCachedQuery';
+import { useInfiniteCachedQuery } from '../../../../hooks/useInfiniteCachedQuery';
 import { stidQueryKeys } from '../../keys/STIDQueryKeys';
 
 interface StidDocumentProps {
